refactor: properties를 불러오는 로직과 config를 통해 client를 제공하는 로직 분리
