--- server/src/main/java/wap/web2/server/config/AwsS3Config.java
@@ -1,38 +1,0 @@
-package wap.web2.server.config;
-
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
-
-@Getter
-@Configuration
-public class AwsS3Config {
-
-    @Value("${aws.accessKeyId}")
-    private String accessKeyId;
-
-    @Value("${aws.secretKey}")
-    private String secretKey;
-
-    @Value("${aws.s3.region}")
-    private String region;
-
-    @Value("${aws.s3.bucketName}")
-    private String bucketName;
-
-    @Bean
-    public S3Client s3Client() {
-        return S3Client.builder()
-                .region(Region.AP_NORTHEAST_2)
-                .credentialsProvider(
-                        StaticCredentialsProvider.create(AwsBasicCredentials.create(accessKeyId, secretKey))
-                )
-                .build();
-    }
-
-}
