+++ server/src/main/java/wap/web2/server/config/S3Config.java
@@ -1,0 +1,34 @@
+package wap.web2.server.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+
+@Configuration
+@RequiredArgsConstructor
+@EnableConfigurationProperties(S3Properties.class)
+public class S3Config {
+
+    private final S3Properties s3Properties;
+
+    @Bean
+    public S3Client s3Client() {
+        return S3Client.builder()
+                .region(Region.of(s3Properties.getS3().getRegion()))
+                .credentialsProvider(
+                        StaticCredentialsProvider.create(
+                                AwsBasicCredentials.create(
+                                        s3Properties.getAccessKeyId(),
+                                        s3Properties.getSecretKey()
+                                )
+                        )
+                )
+                .build();
+    }
+    
+}
